move jena related code to JenaOps
