+++ webapi/src/main/scala/org/knora/webapi/slice/common/ModelOps.scala
@@ -6,18 +6,12 @@
 package org.knora.webapi.slice.common
 
 import org.apache.jena.rdf.model.*
-import org.apache.jena.riot.Lang
-import org.apache.jena.riot.RDFDataMgr
 import org.apache.jena.vocabulary.RDF
 import zio.*
 
-import java.io.ByteArrayInputStream
-import java.io.StringWriter
-import java.nio.charset.StandardCharsets
 import java.time.Instant
 import java.util.UUID
 import scala.jdk.CollectionConverters.*
-import scala.util.Try
 
 import dsp.valueobjects.UuidUtil
 import dsp.valueobjects.UuidUtil.base64Decode
@@ -26,7 +20,6 @@ import org.knora.webapi.messages.OntologyConstants
 import org.knora.webapi.messages.OntologyConstants.KnoraApiV2Complex.ValueCreationDate
 import org.knora.webapi.messages.OntologyConstants.KnoraApiV2Complex.ValueHasUUID
 import org.knora.webapi.messages.SmartIri
-import org.knora.webapi.messages.ValuesValidator
 import org.knora.webapi.slice.admin.domain.model.KnoraProject.Shortcode
 import org.knora.webapi.slice.common.ModelError.IsNoResourceIri
 import org.knora.webapi.slice.common.ModelError.MoreThanOneRootResource
@@ -165,78 +158,3 @@ final case class KnoraApiValueNode(node: RDFNode, belongsTo: Property, shortcode
 }
 
 object KnoraApiValueNode {}
-
-object NodeOps {
-  extension (node: RDFNode) {
-    def getStringLiteral(property: String): Option[String] =
-      node.getStringLiteral(ResourceFactory.createProperty(property))
-    def getStringLiteral(property: Property): Option[String] =
-      node.getLiteral(property).flatMap(lit => Try(lit.getString).toOption)
-    def getLiteral(property: Property): Option[Literal] =
-      node.getObject(property).flatMap(obj => Try(obj.asLiteral()).toOption)
-    def getObject(property: Property): Option[RDFNode] =
-      node.getStatement(property).map(_.getObject)
-    def getStatement(property: Property): Option[Statement] =
-      node.toResourceOption.flatMap(r => Option(r.getProperty(property)))
-    def toResourceOption: Option[Resource] = Try(node.asResource()).toOption
-    def getDateTimeProperty(property: Property): Either[String, Option[Instant]] =
-      node
-        .getLiteral(property)
-        .map { lit =>
-          Right(lit)
-            .filterOrElse(
-              _.getDatatypeURI == OntologyConstants.Xsd.DateTimeStamp,
-              s"Invalid data type (should be xsd:dateTimeStamp) for value: ${lit.getLexicalForm}",
-            )
-            .map(_.getLexicalForm)
-            .flatMap(str =>
-              ValuesValidator
-                .xsdDateTimeStampToInstant(str)
-                .toRight(s"Invalid xsd:dateTimeStamp value: $str")
-                .map(Some(_)),
-            )
-        }
-        .fold(Right(None))(identity)
-  }
-}
-
-object ResourceOps {
-  extension (res: Resource) {
-    def property(p: Property): Option[Statement] = Option(res.getProperty(p))
-    def rdfsType(): Option[String]               = Option(res.getPropertyResourceValue(RDF.`type`)).flatMap(_.uri)
-    def uri: Option[String]                      = Option(res.getURI)
-  }
-}
-object StatementOps {
-  extension (stmt: Statement) {
-    def subjectUri(): Option[String] = Option(stmt.getSubject.getURI)
-    def objectUri(): Option[String]  = Try(stmt.getObject.asResource()).toOption.flatMap(r => Option(r.getURI))
-  }
-}
-
-object ModelOps { self =>
-
-  extension (model: Model) {
-    def printTurtle: UIO[Unit] =
-      ZIO.scoped {
-        for {
-          writer <- ZIO.fromAutoCloseable(ZIO.succeed(new StringWriter()))
-          _      <- ZIO.attempt(RDFDataMgr.write(writer, model, Lang.TURTLE))
-          _      <- Console.printLine(writer.toString)
-        } yield ()
-      }.logError.ignore
-  }
-
-  def fromJsonLd(str: String): ZIO[Scope, ParseError, Model] = from(str, Lang.JSONLD)
-
-  private val createModel =
-    ZIO.acquireRelease(ZIO.succeed(ModelFactory.createDefaultModel()))(m => ZIO.succeed(m.close()))
-
-  def from(str: String, lang: Lang): ZIO[Scope, ParseError, Model] =
-    for {
-      m <- createModel
-      _ <- ZIO
-             .attempt(RDFDataMgr.read(m, ByteArrayInputStream(str.getBytes(StandardCharsets.UTF_8)), lang))
-             .mapError(ModelError.parseError)
-    } yield m
-}
